Split models module into submodules
