+++ feedinlib/models/__init__.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+
+"""
+Feed-in model classes.
+
+SPDX-FileCopyrightText: Birgit Schachler
+SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
+SPDX-FileCopyrightText: Stephan Günther
+SPDX-FileCopyrightText: Stephen Bosch
+SPDX-FileCopyrightText: Patrik Schönfeldt <patrik.schoenfeldt@dlr.de>
+
+SPDX-License-Identifier: MIT
+
+This module provides abstract classes as blueprints for classes that implement
+feed-in models for weather dependent renewable energy resources (in base).
+Furthermore, this module holds implementations of feed-in models (other files).
+"""
+
+from .pvlib import Pvlib
+from .windpowerlib import (WindpowerlibTurbine, WindpowerlibTurbineCluster)
+from .base import get_power_plant_data
+from .geometric_solar import GeometricSolar
