+++ feedinlib/powerplants.py
@@ -3,12 +3,6 @@
 """
 Power plant classes for specific weather dependent renewable energy resources.
 
-Power plant classes act as data holders for the attributes making up a
-power plant's specification. These classes should only contain little logic.
-Computing the actual feed-in provided by a power plant is done by the models
-(see models.py). The model the feed-in is calculated with is specified in
-the `model` attribute.
-
 SPDX-FileCopyrightText: Birgit Schachler
 SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
 SPDX-FileCopyrightText: Stephan Günther
@@ -18,6 +12,11 @@ SPDX-FileCopyrightText: Patrik Schönfeldt <patrik.schoenfeldt@dlr.de>
 
 SPDX-License-Identifier: MIT
 
+Power plant classes act as data holders for the attributes making up a
+power plant's specification. These classes should only contain little logic.
+Computing the actual feed-in provided by a power plant is done by the models
+(see models.py). The model the feed-in is calculated with is specified in
+the `model` attribute.
 """
 
 from abc import ABC, abstractmethod
